lab1/requirements.txt
markdown
flask
google-generativeai
rebuff